requirements.txt
# --- Development tools ---
ipykernel
yapf

# --- Visualization & Geospatial ---
matplotlib
seaborn
pygeohash

# --- Core scientific stack ---
numpy>=1.26.0          # compatible with Python 3.12
pandas>=2.0.0
scipy>=1.11.0          # newer version supports Python 3.12
scikit-learn>=1.3.1
h5py>=3.10.0

# --- Google Cloud ecosystem ---
google-cloud-bigquery
google-cloud-storage>=2.14.0
google-api-core>=2.8.2
googleapis-common-protos>=1.63.0
protobuf>=4.25.0       # 3.x breaks with Python 3.12
pyarrow>=14.0.1
db-dtypes

# --- TensorFlow (auto-detect by platform) ---
tensorflow-macos>=2.15.0; sys_platform == 'darwin' and 'ARM' in platform_version   # Apple Silicon Macs
tensorflow>=2.15.0; sys_platform == 'darwin' and 'ARM' not in platform_version     # Intel Macs
tensorflow>=2.15.0; sys_platform != 'darwin'                                      # Windows & Linux

# --- MLOps & workflow tools ---
mlflow>=2.11.0          # 2.1.1 too old for Python 3.12
prefect>=2.19.2
python-dotenv

# --- Database driver ---
psycopg2-binary>=2.9.9   # version compatible with Python 3.12

# --- Image processing ---
pillow
scikit-image
opencv-python-headless # opencv-python-headless plutôt que opencv-python pour éviter les problèmes de GUI dans Docker.

# --- API Colorizer ---
fastapi
uvicorn[standard] # [standard] --> recommandé officiellement par FastApi (plus de performance et possibilités)
python-multipart  # upload de fichiers dans FastAPI
